CM3: Add resource table definition for use with remoteproc
Add a resource table so that the remoteproc framework within the Linux
kernel is able to load the firmware when the (Linux) driver probes in
order to specify where the needed memory regions are. Also add a TRACE
resource so that remoteproc can access the LOGBUF from the kernel for
debug output.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>