CM3: msg: Add halt flag to param3
Add a flag into param3 IPC register to allow user of the firmware
to set this bit and cause the firmware to wait at a loop at the last
possible point in DS0 to allow for connection by JTAG for debug
purposes. Once this occurs the program counter can be adjusted to exit
the loop and continue to suspend.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>