CM3: msg: Add halt flag to param3
Add a flag into param3 IPC register to allow user of the firmware
to set this bit and cause the firmware to wait at a loop at the last
possible point in all of the suspend states to allow for connection by
JTAG for debug purposes. Once this occurs the program counter can
be adjusted to exit the loop and continue to suspend.

This is useful for inspecting the PRCM or any other registers visible
from the M3 to debug suspend failures.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>