CM3: Fix Add i2c speed value in voltage scaling binaries
An incorrect value of 00 34 is present for the i2c bus speed given in
all of the voltage scaling binaries. This value should be 64 00 to properly
represent 100 kHz due to the little-endianess of the two-byte value. Also,
because of this change, we must add a 00 byte at the end of the sleep
sequence to properly terminate it which then increases the offset to the
wake sequence in all of the binaries by 1.

The correct value of 100 kHz should be used to ensure reliability of the
i2c transaction, although the incorrect value seems to have worked
without visible issues.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>