CM3: ddr: Fix vtp_enable sequence
A typo in the code for vtp_enable when toggling the CLRZ bit in
VTP0_CTRL_REG is accidentally clearing the entire register rather so any
values we program before are erased. This has not caused any observed
issues but may cause issues on certain specific silicon due to improper
VTP compensation.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>